+++ extension/src/test/suite/extension.test.ts
@@ -9,12 +9,24 @@ import { activate } from '../../extension';
 import { stubRemoteName, stubConfig } from '../stubs';
 
 suite('Extension Test Suite', () => {
+	let context: vscode.ExtensionContext;
+
 	beforeEach(() => {
-		delete process.env['DISPLAY'];
+		context = new MockContext();
 	});
 
 	afterEach(() => {
 		sinon.restore();
+		context.subscriptions.forEach((s) => s.dispose());
+	});
+
+	test('No remote', async () => {
+		stubRemoteName(undefined);
+
+		await activate(context);
+
+		assertVariableUndefined(context, 'DISPLAY');
+		assertVariableUndefined(context, 'LIBGL_ALWAYS_INDIRECT');
 	});
 
 	test('Container: attached', async () => {
@@ -24,9 +36,9 @@ suite('Extension Test Suite', () => {
 			screen: 2,
 		});
 
-		await activate();
+		await activate(context);
 
-		assert.equal(process.env['DISPLAY'], 'host.docker.internal:1.2');
+		assertVariableReplaced(context, 'DISPLAY', 'host.docker.internal:1.2');
 	});
 
 	test('Container: dev', async () => {
@@ -36,20 +48,20 @@ suite('Extension Test Suite', () => {
 			screen: 2,
 		});
 
-		await activate();
+		await activate(context);
 
-		assert.equal(process.env['DISPLAY'], 'host.docker.internal:1.2');
+		assertVariableReplaced(context, 'DISPLAY', 'host.docker.internal:1.2');
 	});
 
 	test('Container: disabled', async () => {
 		stubRemoteName('attached-container');
 		stubConfig({
-			'container.enabled': false,
+			'container.enable': false,
 		});
 
-		await activate();
+		await activate(context);
 
-		assert.equal(process.env['DISPLAY'], undefined);
+		assertVariableUndefined(context, 'DISPLAY');
 	});
 
 	test('WSL: enabled', async () => {
@@ -59,20 +71,20 @@ suite('Extension Test Suite', () => {
 			screen: 2,
 		});
 
-		await activate();
+		await activate(context);
 
-		assert.equal(process.env['DISPLAY'], 'localhost:1.2');
+		assertVariableReplaced(context, 'DISPLAY', 'localhost:1.2');
 	});
 
 	test('WSL: disabled', async () => {
 		stubRemoteName('wsl');
 		stubConfig({
-			'WSL.enabled': false,
+			'WSL.enable': false,
 		});
 
-		await activate();
+		await activate(context);
 
-		assert.equal(process.env['DISPLAY'], undefined);
+		assertVariableUndefined(context, 'DISPLAY');
 	});
 
 	test('SSH: enabled', async () => {
@@ -101,23 +113,151 @@ suite('Extension Test Suite', () => {
 			})
 			.resolves('localhost:1.2');
 
-		await activate();
+		await activate(context);
 
 		assert(stub.called);
-		assert.equal(process.env['DISPLAY'], 'localhost:1.2');
+		assertVariableReplaced(context, 'DISPLAY', 'localhost:1.2');
 	});
 
 	test('SSH: disabled', async () => {
 		stubRemoteName('ssh-remote');
 		stubConfig({
-			'SSH.enabled': false,
+			'SSH.enable': false,
 		});
 
 		const spy = sinon.spy(vscode.commands, 'executeCommand');
 
-		await activate();
+		await activate(context);
 
 		assert(spy.neverCalledWithMatch('remote-x11-ssh.connect'));
-		assert.equal(process.env['DISPLAY'], undefined);
+		assertVariableUndefined(context, 'DISPLAY');
+	});
+
+	test('extraVariables', async () => {
+		stubRemoteName('wsl');
+		stubConfig({
+			extraVariables: ['FOO=1', 'BAR = 2', ' BAZ  =3  '],
+		});
+
+		await activate(context);
+
+		assertVariableReplaced(context, 'FOO', '1');
+		assertVariableReplaced(context, 'BAR', '2');
+		assertVariableReplaced(context, 'BAZ', '3');
+	});
+
+	test('extraVariables: default', async () => {
+		stubRemoteName('wsl');
+		stubConfig({});
+
+		await activate(context);
+
+		assertVariableReplaced(context, 'LIBGL_ALWAYS_INDIRECT', '1');
+	});
+
+	test('extraVariables: empty', async () => {
+		stubRemoteName('wsl');
+		stubConfig({
+			extraVariables: [],
+		});
+
+		await activate(context);
+
+		assertVariableUndefined(context, 'LIBGL_ALWAYS_INDIRECT');
 	});
 });
+
+function assertVariableReplaced(context: vscode.ExtensionContext, variable: string, value: string) {
+	assert.deepStrictEqual(context.environmentVariableCollection.get(variable), {
+		type: vscode.EnvironmentVariableMutatorType.Replace,
+		value,
+	});
+}
+
+function assertVariableUndefined(context: vscode.ExtensionContext, variable: string) {
+	assert.strictEqual(context.environmentVariableCollection.get(variable), undefined);
+}
+
+class MockContext implements vscode.ExtensionContext {
+	public readonly subscriptions: { dispose(): any }[] = [];
+
+	public readonly workspaceState = new MockMemento();
+	public readonly globalState = new MockMemento();
+
+	public readonly extensionUri = vscode.Uri.file('/fake/path');
+	public get extensionPath() {
+		return this.extensionUri.fsPath;
+	}
+
+	public readonly environmentVariableCollection = new MockEnvironmentVariableCollection();
+
+	public asAbsolutePath(relativePath: string) {
+		return vscode.Uri.joinPath(this.extensionUri, relativePath).fsPath;
+	}
+
+	public readonly storagePath = undefined;
+	public readonly globalStoragePath = '/fake/path';
+	public readonly logPath = '/fake/path';
+}
+
+class MockMemento implements vscode.Memento {
+	public get<T>(_key: string, _defaultValue?: T) {
+		return undefined;
+	}
+
+	public update(_key: string, _value: any) {
+		return Promise.resolve();
+	}
+}
+
+class MockEnvironmentVariableCollection implements vscode.EnvironmentVariableCollection {
+	public persistent = true;
+
+	private map = new Map<string, vscode.EnvironmentVariableMutator>();
+
+	public replace(variable: string, value: string): void {
+		this.map.set(variable, {
+			type: vscode.EnvironmentVariableMutatorType.Replace,
+			value,
+		});
+	}
+
+	public append(variable: string, value: string): void {
+		this.map.set(variable, {
+			type: vscode.EnvironmentVariableMutatorType.Append,
+			value,
+		});
+	}
+
+	public prepend(variable: string, value: string): void {
+		this.map.set(variable, {
+			type: vscode.EnvironmentVariableMutatorType.Prepend,
+			value,
+		});
+	}
+
+	public get(variable: string): vscode.EnvironmentVariableMutator | undefined {
+		return this.map.get(variable);
+	}
+
+	public forEach(
+		callback: (
+			variable: string,
+			mutator: vscode.EnvironmentVariableMutator,
+			collection: vscode.EnvironmentVariableCollection,
+		) => any,
+		thisArg?: any,
+	): void {
+		for (const [variable, mutator] of this.map) {
+			callback.apply(thisArg, [variable, mutator, this]);
+		}
+	}
+
+	public delete(variable: string): void {
+		this.map.delete(variable);
+	}
+
+	public clear(): void {
+		this.map.clear();
+	}
+}